Adding more elements.
Cython code integrated in CMake. .so libraries generated
for the Cython + cpp code. Also added tests in the cpp part (not in the python
modules generated).

Added a little web.py application that uses the generated module. The intention
is to show the integration of cpp code into a python web app via Cython.

Still some manual steps, like the installation of the python modules.

The only Python library used is web.py